+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Only write entries that are worth mentioning to users.
 - Lib: Add `thinking: bool` parameter to `ShellApp.run` method
 - Lib: Rename `KimiSoul.set_thinking_mode` to `KimiSoul.set_thinking`
 - UI: Better checks and notices for LLM model capabilities
+- UI: Clear the screen for `/clear` meta command
 
 ## [0.50] - 2025-11-07